statement-service/src/test/java/ru/bank/jd/junit/service/StatementServiceTest.java
package ru.bank.jd.junit.service;

import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import ru.bank.jd.component.RequestDealServiceRest;
import ru.bank.jd.dto.LoanOfferDto;
import ru.bank.jd.dto.LoanStatementRequestDto;
import ru.bank.jd.service.StatementService;
import java.math.BigDecimal;
import java.time.LocalDate;
import java.util.List;
import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
import static org.junit.jupiter.api.Assertions.assertEquals;
import static org.mockito.ArgumentMatchers.any;
import static org.mockito.Mockito.when;

@ExtendWith(MockitoExtension.class)
class StatementServiceTest {

    @Mock
    private RequestDealServiceRest dealServiceRest;

    @InjectMocks
    private StatementService statementService;


    @Test
    void searchOfferTest() {
        when(dealServiceRest.callCalculateOffer(any()))
                .thenReturn(List.of(new LoanOfferDto(), new LoanOfferDto(), new LoanOfferDto(), new LoanOfferDto()));

        assertEquals(4, statementService.searchOffer(getLoanStatementRequestDto()).size());
    }

    @Test
    void selectOfferTest() {
        assertDoesNotThrow(() -> statementService.selectOffer(new LoanOfferDto()));
    }

    private LoanStatementRequestDto getLoanStatementRequestDto() {
        return LoanStatementRequestDto.builder()
                .amount(new BigDecimal("100000"))
                .term(12)
                .firstName("Иван")
                .lastName("Иванов")
                .middleName("Иванович")
                .email("ivanov@example.com")
                .birthdate(LocalDate.of(1990, 5, 15))
                .passportSeries("1234")
                .passportNumber("567890")
                .build();
    }
}